rename of channel layout, move to 64 bits and added raw audio in ISOBMFF 23003-5

+++ src/filters/ff_common.h
@@ -69,6 +69,6 @@ u32 ffmpeg_stream_type_to_gpac(u32 streamtype);
 void ffmpeg_set_enc_dec_flags(const AVDictionary *options, AVCodecContext *ctx);
 void ffmpeg_set_mx_dmx_flags(const AVDictionary *options, AVFormatContext *ctx);
 
-u32 ffmpeg_channel_layout_from_gpac(u32 gpac_ch_layout);
-u32 ffmpeg_channel_layout_to_gpac(u32 ff_ch_layout);
+u64 ffmpeg_channel_layout_from_gpac(u64 gpac_ch_layout);
+u64 ffmpeg_channel_layout_to_gpac(u64 ff_ch_layout);
